requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# cogs\DerpiBooru.py: 3
# cogs\Reddit.py: 4
# cogs\e926.py: 3
aiohttp == 2.2.5

# fluffy.py: 9,10
cogs == 0.4.1

# cogs\DerpiBooru.py: 4
# cogs\HelpModule.py: 1,2
# cogs\Reddit.py: 5
# cogs\e926.py: 4,5
# cogs\Fun_commands.py: 3
# cogs\googlemodule.py: 1
# cogs\hourlyfox.py: 1,2
# cogs\tags.py: 1
# cogs\useful_commands.py: 1
# cogs\wikipedia.py: 1,3
# fluffy.py: 6,7
discord.py == 1.0.0a1322

# cogs\TwitterManager.py: 1
peony_twitter == 1.0.4

# cogs\googlemodule.py: 5
protobuf == 3.5.2.post1

# cogs\core\configJSON.py: 1
ujson == 1.35

# cogs\wikipedia.py: 2
wikipedia == 1.4.0

# cogs\googlemodule.py: 5
google==2.0.1

# cogs\googlemodule.py: 3
google-search==1.0.2

